devoir-2/pyproject.toml
[project]
name = "devoir-2"
version = "0.1.0"
description = "IFT3700 - Devoir 2"
readme = "README.md"
requires-python = "==3.9.*"
dependencies = [
    "black>=24.8.0",
    "ffmpeg-python>=0.2.0",
    "jupyterlab>=4.2.5",
    "librosa>=0.10.2.post1",
    "matplotlib>=3.9.2",
    "numpy>=2.0.2",
    "pandas>=2.2.2",
    "python-dotenv>=1.0.1",
    "scipy>=1.13.1",
    "seaborn>=0.13.2",
    "soundfile>=0.12.1",
    "tqdm>=4.66.5",
    "yt-dlp>=2024.9.27",
]
